Fixed DWORD type

+++ bitTypes.h
@@ -1,4 +1,4 @@
 #pragma once
 
 typedef unsigned short int WORD;
-typedef unsigned long int DWORD;
+typedef unsigned int DWORD;
